+++ lib/proof/data/repositories/proof_repository_impl.dart
@@ -98,8 +98,8 @@ class ProofRepositoryImpl extends ProofRepository {
   @override
   Future<Uint8List> calculateAtomicQueryInputs(
       {required String id,
-      required int profileNonce,
-      required int claimSubjectProfileNonce,
+      required BigInt profileNonce,
+      required BigInt claimSubjectProfileNonce,
       required ClaimEntity claim,
       required ProofScopeRequest request,
       ProofEntity? incProof,
